commit 18/01/2018

+++ Makefile
@@ -14,7 +14,8 @@ diffall:
 case:
 	~/.dotfile/command/ctest.py ${file} diff case ${number}
 clean:
-	rm a.out
+	[ -f a.out ] && rm a.out || echo "a.out not found"
+	[ -f Main.class ] && Main.class || echo "main.class not found"
 move:
 	mv *.cpp finish-problem/
 mv:
